[MERGE] Merge CS 1.5 back to master

+++ core/server/worker/src/main/java/alluxio/worker/netty/DataServerUfsFileWriteHandler.java
@@ -16,12 +16,13 @@ import alluxio.network.protocol.RPCProtoMessage;
 import alluxio.proto.dataserver.Protocol;
 import alluxio.security.authorization.Mode;
 import alluxio.underfs.UfsManager;
-import alluxio.underfs.UfsManager.Ufs;
+import alluxio.underfs.UfsManager.UfsInfo;
 import alluxio.underfs.UnderFileSystem;
 import alluxio.underfs.options.CreateOptions;
 
 import com.codahale.metrics.Counter;
 import io.netty.buffer.ByteBuf;
+import io.netty.channel.Channel;
 
 import java.io.IOException;
 import java.io.OutputStream;
@@ -46,35 +47,37 @@ final class DataServerUfsFileWriteHandler extends DataServerWriteHandler {
 
   private class FileWriteRequestInternal extends WriteRequestInternal {
     private final String mUfsPath;
-    private final UnderFileSystem mUnderFileSystem;
-    private final OutputStream mOutputStream;
-    private final Counter mCounter;
+    private final Protocol.WriteRequest mWriteRequest;
+
+    private UnderFileSystem mUnderFileSystem;
+    private OutputStream mOutputStream;
+    private Counter mCounter;
 
     FileWriteRequestInternal(Protocol.WriteRequest request) throws Exception {
       super(request.getId());
-      Protocol.CreateUfsFileOptions createUfsFileOptions = request.getCreateUfsFileOptions();
-      mUfsPath = createUfsFileOptions.getUfsPath();
-      Ufs ufs = mUfsManager.get(createUfsFileOptions.getMountId());
-      mUnderFileSystem = ufs.getUfs();
-      mOutputStream = mUnderFileSystem.create(mUfsPath,
-          CreateOptions.defaults().setOwner(createUfsFileOptions.getOwner())
-              .setGroup(createUfsFileOptions.getGroup())
-              .setMode(new Mode((short) createUfsFileOptions.getMode())));
-      String ufsName = MetricsSystem.escapeURI(ufs.getUfsMountPointUri());
-      String metricName = String.format("BytesWrittenUfs-Ufs:%s", ufsName);
-      mCounter = MetricsSystem.workerCounter(metricName);
+      mWriteRequest = request;
+      mUfsPath = request.getCreateUfsFileOptions().getUfsPath();
     }
 
     @Override
-    public void close() throws IOException {
-      mOutputStream.close();
+    public void close(Channel channel) throws IOException {
+      if (mOutputStream == null) {
+        createUfsFile(channel);
+      }
+      if (mOutputStream != null) {
+        mOutputStream.close();
+        mOutputStream = null;
+      }
     }
 
     @Override
     void cancel() throws IOException {
       // TODO(calvin): Consider adding cancel to the ufs stream api.
-      mOutputStream.close();
-      mUnderFileSystem.deleteFile(mUfsPath);
+      if (mOutputStream != null && mUnderFileSystem != null) {
+        mOutputStream.close();
+        mUnderFileSystem.deleteFile(mUfsPath);
+        mOutputStream = null;
+      }
     }
 
     @Override
@@ -117,12 +120,48 @@ final class DataServerUfsFileWriteHandler extends DataServerWriteHandler {
   }
 
   @Override
-  protected void writeBuf(ByteBuf buf, long pos) throws Exception {
+  protected void writeBuf(Channel channel, ByteBuf buf, long pos) throws Exception {
+    FileWriteRequestInternal request = (FileWriteRequestInternal) mRequest;
+    if (request.mOutputStream == null) {
+      createUfsFile(channel);
+    }
+
     buf.readBytes(((FileWriteRequestInternal) mRequest).mOutputStream, buf.readableBytes());
   }
 
   @Override
   protected void incrementMetrics(long bytesWritten) {
     ((FileWriteRequestInternal) mRequest).mCounter.inc(bytesWritten);
+  }
+
+  private void createUfsFile(Channel channel) throws IOException {
+    FileWriteRequestInternal request = (FileWriteRequestInternal) mRequest;
+    Protocol.CreateUfsFileOptions createUfsFileOptions =
+        request.mWriteRequest.getCreateUfsFileOptions();
+    // ALLUXIO CS ADD
+    // Before interacting with the UFS manager, make sure the user is set.
+    String user = channel.attr(alluxio.netty.NettyAttributes.CHANNEL_KERBEROS_USER_KEY).get();
+    if (user != null) {
+      alluxio.security.authentication.AuthenticatedClientUser.set(user);
+    }
+    // ALLUXIO CS END
+    UfsInfo ufsInfo = mUfsManager.get(createUfsFileOptions.getMountId());
+    request.mUnderFileSystem = ufsInfo.getUfs();
+    request.mOutputStream = request.mUnderFileSystem.create(request.mUfsPath,
+        CreateOptions.defaults().setOwner(createUfsFileOptions.getOwner())
+            .setGroup(createUfsFileOptions.getGroup())
+            .setMode(new Mode((short) createUfsFileOptions.getMode())));
+    String ufsString = MetricsSystem.escape(ufsInfo.getUfsMountPointUri());
+    // ALLUXIO CS REPLACE
+    // String metricName = String.format("BytesWrittenUfs-Ufs:%s", ufsString);
+    // ALLUXIO CS WITH
+    String metricName;
+    if (user == null) {
+      metricName = String.format("BytesWrittenUfs-Ufs:%s", ufsString);
+    } else {
+      metricName = String.format("BytesWrittenUfs-Ufs:%s-User:%s", ufsString, user);
+    }
+    // ALLUXIO CS END
+    request.mCounter = MetricsSystem.workerCounter(metricName);
   }
 }
